refactor: data api Service분리
